Adding G2U_ImageChunk message and ability to handle (and display) it from webots_keyboard_controller

+++ include/anki/cozmo/game/comms/gameMessageHandler.h
@@ -92,7 +92,7 @@ namespace Cozmo {
     // message ID.
     struct {
       u8 priority;
-      u8 size;
+      u16 size;
       Result (GameMessageHandler::*ProcessPacketAs)(const u8* );
     } lookupTable_[NUM_UI_MSG_IDS+1] = {
       {0, 0, 0}, // Empty entry for NO_MESSAGE_ID
